pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
classifiers = [
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
]
name = "nomad-forematics"
description = "NOMAD plugin for forematics experiments"
version = "0.1.3"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    { name = "Miquel Casademont", email = "mcasademont@icmab.es" },
]
maintainers = [
    { name = "Miquel Casademont", email = "mcasademont@icmab.es" },
]
license = { file = "LICENSE" }
dependencies = [
    "nomad-lab>=1.3.0",
    "python-magic-bin; sys_platform == 'win32'",
]

[project.urls]
Repository = "https://github.com/mcasademont9/nomad-forematics"

[project.optional-dependencies]
dev = ["ruff", "pytest", "structlog"]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 88 #88 was the default value
indent-width = 4


[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # isort
    "I",
    # pylint
    "PL",
]

ignore = [
    "F403", # 'from module import *' used; unable to detect undefined names
    "E501", # To ignore "line too long" errors
]

fixable = ["ALL"]

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# this is entirely optional, you can remove this if you wish to
[tool.ruff.format]
# use single quotes for strings.
quote-style = "single"

# indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

#[tool.setuptools]
#packages = ["nomad_forematics"]
#package-dir = { "" = "src" }
#exclude = [".venv"]

#[tool.setuptools.packages.find]
#where = ["src"]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
include = ["nomad_forematics"]
exclude = [".pyenv"]

[project.entry-points.'nomad.plugin']
parser_entry_point = "nomad_forematics.parsers:parser_entry_point"
#schema_package_entry_point = "nomad_forematics.schema_packages:schema_package_entry_point"
substrate = "nomad_forematics.schema_packages:substrate"
solution = "nomad_forematics.schema_packages:solution"
experiment = "nomad_forematics.schema_packages:experiment"
processing = "nomad_forematics.schema_packages:processing"
normalizer_entry_point = "nomad_forematics.normalizers:normalizer_entry_point"
app_entry_point = "nomad_forematics.apps:app_entry_point"
example_upload_entry_point = "nomad_forematics.example_uploads:example_upload_entry_point"
[tool.cruft]

# Avoid updating workflow files, this leads to permissions issues
skip = [".github/*"]
